+++ src/containers/Form/Skills/__tests__/index.test.js
@@ -15,12 +15,15 @@ import {
   userHasUpdatePermission
 } from '../../../../redux/modules/entities/selectors';
 import { selectFormInitialValues } from '../../../../redux/modules/form/selectors';
+import { getHasProficiencyFormValue } from '../../../../redux/modules/entities/skills/selectors';
 
 jest.mock('../../../../redux/modules/entities/selectors');
 jest.mock('../../../../redux/modules/form/selectors');
+jest.mock('../../../../redux/modules/entities/skills/selectors');
 getCurrentForm.mockImplementation(() => 'gets form from state');
 getSelectedEntityId.mockImplementation(() => 'mockId');
 isInherited.mockImplementation(() => false);
+getHasProficiencyFormValue.mockImplementation(() => true);
 isCreating.mockImplementation(() => true);
 userHasUpdatePermission.mockImplementation(() => true);
 selectFormInitialValues.mockImplementation(