+++ src/redux/modules/entities/epics.js
@@ -12,14 +12,17 @@ import { fromPromise } from 'rxjs/observable/fromPromise';
 import { from } from 'rxjs/observable/from';
 import { forkJoin } from 'rxjs/observable/forkJoin';
 import { Toast } from 'cx-ui-components';
-import { clearFields, touch } from 'redux-form';
+import { clearFields, touch, change } from 'redux-form';
 
 import * as MODALS from '../../../containers/ConfirmationDialog/constants.js';
 
 import { sdkPromise } from '../../../utils/sdk';
 import { handleError, handleSuccess, handleBulkSuccess } from './handleResult';
 
-import { entityAddedToList, entityRemovedFromList } from '../../modules/entities/listItemSelectors';
+import {
+  entityAddedToList,
+  entityRemovedFromList
+} from '../../modules/entities/listItemSelectors';
 
 import { uploadCsv, setEntityUpdating } from './index';
 import { isInIframe } from 'serenova-js-utils/browser';
@@ -35,8 +38,11 @@ import {
   getConfirmationDialogType,
   getConfirmationDialogMetaData,
   getSelectedEntityName,
-  getSelectedEntityBulkChangeItems
+  getSelectedEntityBulkChangeItems,
+  getSelectedEntityFormId
 } from './selectors';
+
+import { getHasProficiencyFormValue } from './skills/selectors';
 
 import { entitiesMetaData } from './metaData';
 
@@ -119,6 +125,17 @@ export const FocusOutboundIdentifiersValueFormField = (action$, store) =>
     )
     .map(a => touch(a.meta.form, 'value'));
 
+export const ToggleHasProficiencyFormField = (action$, store) =>
+  action$
+    .ofType('TOGGLE_PROFICIENCY')
+    .map(a =>
+      change(
+        getSelectedEntityFormId(store.getState()),
+        'hasProficiency',
+        !getHasProficiencyFormValue(store.getState())
+      )
+    );
+
 export const FormSubmission = (action$, store) =>
   action$
     .ofType('FORM_SUBMIT')
@@ -153,7 +170,11 @@ export const FormSubmission = (action$, store) =>
 
 export const FetchData = action$ =>
   action$.ofType('FETCH_DATA').mergeMap(a =>
-    fromPromise(sdkPromise(entitiesMetaData[a.entityName].entityApiRequest('get', 'mainEntity')))
+    fromPromise(
+      sdkPromise(
+        entitiesMetaData[a.entityName].entityApiRequest('get', 'mainEntity')
+      )
+    )
       .map(response => handleSuccess(response, a))
       .catch(error => handleError(error, a))
   );
@@ -163,7 +184,10 @@ export const FetchDataItem = action$ =>
     .ofType('FETCH_DATA_ITEM')
     .debounceTime(300)
     .map(a => {
-      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest('get', 'singleMainEntity');
+      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest(
+        'get',
+        'singleMainEntity'
+      );
       a.sdkCall.data = { [removeLastLetter(a.entityName) + 'Id']: a.id };
       return { ...a };
     })
@@ -175,7 +199,11 @@ export const FetchDataItem = action$ =>
 
 export const getTenantPermissions = action$ =>
   action$
-    .ofType('SET_CURRENT_ENTITY', 'START_SUPERVISOR_TOOLBAR_$', 'FETCH_BRANDING_$')
+    .ofType(
+      'SET_CURRENT_ENTITY',
+      'START_SUPERVISOR_TOOLBAR_$',
+      'FETCH_BRANDING_$'
+    )
     .filter(a => !isInIframe())
     .switchMap(a =>
       fromPromise(
@@ -201,7 +229,10 @@ export const CreateEntity = (action$, store) =>
     .ofType('CREATE_ENTITY')
     .filter(({ entityName }) => hasCustomCreateEntity(entityName))
     .map(a => {
-      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest('create', 'singleMainEntity');
+      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest(
+        'create',
+        'singleMainEntity'
+      );
       a.sdkCall.data = a.values;
       return { ...a };
     })
@@ -211,24 +242,30 @@ export const CreateEntity = (action$, store) =>
           handleSuccess(
             response,
             a,
-            `${camelCaseToRegularFormAndRemoveLastLetter(a.entityName)} was created successfully!`
+            `${camelCaseToRegularFormAndRemoveLastLetter(
+              a.entityName
+            )} was created successfully!`
           )
         )
         .catch(error => handleError(error, a))
     );
 
 export const CreateEntityFullfilled = action$ =>
-  action$.ofType('CREATE_ENTITY_FULFILLED', 'COPY_CURRENT_ENTITY_FULFILLED').map(a => ({
-    type: 'SET_SELECTED_ENTITY_ID',
-    entityId: a.response.result.id || a.response.result.userId
-  }));
+  action$
+    .ofType('CREATE_ENTITY_FULFILLED', 'COPY_CURRENT_ENTITY_FULFILLED')
+    .map(a => ({
+      type: 'SET_SELECTED_ENTITY_ID',
+      entityId: a.response.result.id || a.response.result.userId
+    }));
 
 export const CopyEntity = (action$, store) =>
   action$
     .ofType('COPY_CURRENT_ENTITY')
     .map(a => {
       a.entityName = getCurrentEntity(store.getState());
-      a.sdkCall = entitiesMetaData[getCurrentEntity(store.getState())].entityApiRequest('create', 'singleMainEntity');
+      a.sdkCall = entitiesMetaData[
+        getCurrentEntity(store.getState())
+      ].entityApiRequest('create', 'singleMainEntity');
       a.sdkCall.data = getSelectedEntity(store.getState()).toJS();
       a.sdkCall.data.name += ' (Copy)';
       return { ...a };
@@ -239,7 +276,9 @@ export const CopyEntity = (action$, store) =>
           handleSuccess(
             response,
             a,
-            `${camelCaseToRegularFormAndRemoveLastLetter(a.entityName)} was copied successfully!`
+            `${camelCaseToRegularFormAndRemoveLastLetter(
+              a.entityName
+            )} was copied successfully!`
           )
         )
         .catch(error => handleError(error, a))
@@ -250,7 +289,10 @@ export const UpdateEntity = (action$, store) =>
     .ofType('UPDATE_ENTITY')
     .filter(({ entityName }) => hasCustomUpdateEntity(entityName))
     .map(a => {
-      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest('update', 'singleMainEntity');
+      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest(
+        'update',
+        'singleMainEntity'
+      );
       a.sdkCall.data = {
         ...a.values,
         [removeLastLetter(a.entityName) + 'Id']: a.entityId
@@ -263,7 +305,9 @@ export const UpdateEntity = (action$, store) =>
           handleSuccess(
             response,
             a,
-            `${camelCaseToRegularFormAndRemoveLastLetter(a.entityName)} was updated successfully!`
+            `${camelCaseToRegularFormAndRemoveLastLetter(
+              a.entityName
+            )} was updated successfully!`
           )
         )
         .catch(error => handleError(error, a))
@@ -274,7 +318,10 @@ export const BulkEntityUpdate = (action$, store) =>
     .ofType('BULK_ENTITY_UPDATE')
     .map(a => {
       a.allIdsToProcess = getSelectedEntityBulkChangeItems(store.getState());
-      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest('update', 'singleMainEntity');
+      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest(
+        'update',
+        'singleMainEntity'
+      );
       a.allSdkCalls = [...a.allIdsToProcess.toJS()].map(item => ({
         ...a.sdkCall,
         data: {
@@ -296,7 +343,9 @@ export const BulkEntityUpdate = (action$, store) =>
         )
       )
         .do(allResult => handleBulkSuccess(allResult))
-        .mergeMap(result => from(result).map(response => handleSuccess(response, a)))
+        .mergeMap(result =>
+          from(result).map(response => handleSuccess(response, a))
+        )
     );
 
 export const ToggleEntity = (action$, store) =>
@@ -310,7 +359,10 @@ export const ToggleEntity = (action$, store) =>
     }))
     .filter(a => a.entityName !== 'users')
     .map(a => {
-      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest('update', 'singleMainEntity');
+      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest(
+        'update',
+        'singleMainEntity'
+      );
       a.sdkCall.data = {
         active: !a.entityStatusActive,
         [removeLastLetter(a.entityName) + 'Id']: a.selectedEntityId
@@ -339,9 +391,13 @@ export const ToggleEntityListItem = (action$, store) =>
       entityName: getCurrentEntity(store.getState())
     }))
     .map(a => {
-      a.sdkCall = entitiesMetaData[a.entityName].entityListItemApiRequest('update');
+      a.sdkCall = entitiesMetaData[a.entityName].entityListItemApiRequest(
+        'update'
+      );
       a.sdkCall.data = {
-        [`${removeLastLetter(entitiesMetaData[a.entityName].dependentEntity)}Id`]: a.entity.id,
+        [`${removeLastLetter(
+          entitiesMetaData[a.entityName].dependentEntity
+        )}Id`]: a.entity.id,
         active: !a.entity.active
       };
       return { ...a };
@@ -360,12 +416,20 @@ export const RemoveListItem = (action$, store) =>
       entityName: getCurrentEntity(store.getState()),
       listId: getSelectedEntityId(store.getState())
     }))
-    .filter(a => a.entityName !== 'roles' && a.entityName !== 'dataAccessReports')
+    .filter(
+      a => a.entityName !== 'roles' && a.entityName !== 'dataAccessReports'
+    )
     .map(a => {
-      a.sdkCall = entitiesMetaData[a.entityName].entityListItemApiRequest('remove');
+      a.sdkCall = entitiesMetaData[a.entityName].entityListItemApiRequest(
+        'remove'
+      );
       a.sdkCall.data = {
-        [`${removeLastLetter(entitiesMetaData[a.entityName].entityName)}Id`]: a.listId,
-        [`${removeLastLetter(entitiesMetaData[a.entityName].dependentEntity)}Id`]: a.listItemId
+        [`${removeLastLetter(
+          entitiesMetaData[a.entityName].entityName
+        )}Id`]: a.listId,
+        [`${removeLastLetter(
+          entitiesMetaData[a.entityName].dependentEntity
+        )}Id`]: a.listItemId
       };
       return { ...a };
     })
@@ -384,12 +448,20 @@ export const AddListItem = (action$, store) =>
       entityName: getCurrentEntity(store.getState()),
       listId: getSelectedEntityId(store.getState())
     }))
-    .filter(a => a.entityName !== 'roles' && a.entityName !== 'dataAccessReports')
+    .filter(
+      a => a.entityName !== 'roles' && a.entityName !== 'dataAccessReports'
+    )
     .map(a => {
-      a.sdkCall = entitiesMetaData[a.entityName].entityListItemApiRequest('add');
+      a.sdkCall = entitiesMetaData[a.entityName].entityListItemApiRequest(
+        'add'
+      );
       a.sdkCall.data = {
-        [`${removeLastLetter(entitiesMetaData[a.entityName].entityName)}Id`]: a.listId,
-        [`${removeLastLetter(entitiesMetaData[a.entityName].dependentEntity)}Id`]: a.listItemId
+        [`${removeLastLetter(
+          entitiesMetaData[a.entityName].entityName
+        )}Id`]: a.listId,
+        [`${removeLastLetter(
+          entitiesMetaData[a.entityName].dependentEntity
+        )}Id`]: a.listItemId
       };
       return a;
     })
@@ -408,14 +480,14 @@ export const AddingListItems = (action$, store) =>
       entityName: getCurrentEntity(store.getState()),
       entityId: getSelectedEntityId(store.getState())
     }))
-    .filter(a => a.entityName === 'roles' || a.entityName === 'dataAccessReports')
+    .filter(
+      a => a.entityName === 'roles' || a.entityName === 'dataAccessReports'
+    )
     .map(a => ({
       ...a,
       values: {
-        [entitiesMetaData[getCurrentEntity(store.getState())].dependentEntity]: entityAddedToList(
-          store.getState(),
-          a.listItemId
-        )
+        [entitiesMetaData[getCurrentEntity(store.getState())]
+          .dependentEntity]: entityAddedToList(store.getState(), a.listItemId)
       }
     }))
     .map(a => ({
@@ -460,7 +532,9 @@ export const addAndRemoveListItemEntities = (action$, store) =>
           handleSuccess(
             response,
             a,
-            `${camelCaseToRegularFormAndRemoveLastLetter(a.entityName)} was updated successfully!`
+            `${camelCaseToRegularFormAndRemoveLastLetter(
+              a.entityName
+            )} was updated successfully!`
           )
         )
         .catch(error => handleError(error, a))
@@ -479,7 +553,11 @@ export const fetchEntityListItems = (action$, store) =>
     .map(a => ({
       ...a,
       values: {
-        path: [a.entityName, a.entityId, camelCaseToKebabCase(a.associatedEntityName)]
+        path: [
+          a.entityName,
+          a.entityId,
+          camelCaseToKebabCase(a.associatedEntityName)
+        ]
       }
     }))
     .concatMap(a =>
@@ -509,11 +587,14 @@ export const RemovingListItems = (action$, store) =>
       entityName: getCurrentEntity(store.getState()),
       entityId: getSelectedEntityId(store.getState())
     }))
-    .filter(a => a.entityName === 'roles' || a.entityName === 'dataAccessReports')
+    .filter(
+      a => a.entityName === 'roles' || a.entityName === 'dataAccessReports'
+    )
     .map(a => ({
       ...a,
       values: {
-        [entitiesMetaData[getCurrentEntity(store.getState())].dependentEntity]: entityRemovedFromList(
+        [entitiesMetaData[getCurrentEntity(store.getState())]
+          .dependentEntity]: entityRemovedFromList(
           store.getState(),
           a.listItemId
         )
@@ -533,7 +614,8 @@ export const FetchFormMetaData = (action$, store) =>
       ...a,
       currentEntityName: getCurrentEntity(store.getState()),
       entityId: getSelectedEntityId(store.getState()),
-      isDefined: name => store.getState().getIn(['Entities', name, 'data']).size === 0
+      isDefined: name =>
+        store.getState().getIn(['Entities', name, 'data']).size === 0
     }))
     .switchMap(
       a =>
@@ -554,7 +636,12 @@ export const FetchSidePanelData = (action$, store) =>
       currentEntityName: getCurrentEntity(store.getState()),
       entityId: getSelectedEntityId(store.getState())
     }))
-    .filter(a => a.entityId !== 'create' && a.entityId !== '' && entitiesMetaData[a.currentEntityName].dependentEntity)
+    .filter(
+      a =>
+        a.entityId !== 'create' &&
+        a.entityId !== '' &&
+        entitiesMetaData[a.currentEntityName].dependentEntity
+    )
     .map(a => ({
       type: 'FETCH_DATA_ITEM',
       entityName: a.currentEntityName,
@@ -601,7 +688,10 @@ export const CreateSubEntity = (action$, store) =>
       entityId: getSelectedEntityId(store.getState())
     }))
     .map(a => {
-      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest('create', 'subEntity');
+      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest(
+        'create',
+        'subEntity'
+      );
       a.sdkCall.data = {
         [`${removeLastLetter(a.entityName)}Id`]: a.entityId,
         itemValue: a.values
@@ -611,7 +701,13 @@ export const CreateSubEntity = (action$, store) =>
     .concatMap(a =>
       fromPromise(sdkPromise(a.sdkCall))
         .map(response =>
-          handleSuccess(response, a, `<i>${camelCaseToRegularForm(a.entityName)}</i> was created successfully!`)
+          handleSuccess(
+            response,
+            a,
+            `<i>${camelCaseToRegularForm(
+              a.entityName
+            )}</i> was created successfully!`
+          )
         )
         .catch(error => handleError(error, a))
     );
@@ -627,7 +723,10 @@ export const UpdateSubEntity = (action$, store) =>
       entityId: getSelectedEntityId(store.getState())
     }))
     .map(a => {
-      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest('update', 'subEntity');
+      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest(
+        'update',
+        'subEntity'
+      );
       a.sdkCall.data = {
         listId: a.selectedEntity.get('id'),
         listItemKey: a.subEntityId,
@@ -638,7 +737,13 @@ export const UpdateSubEntity = (action$, store) =>
     .concatMap(a =>
       fromPromise(sdkPromise(a.sdkCall))
         .map(response =>
-          handleSuccess(response, a, `<i>${camelCaseToRegularForm(a.subEntityName)} </i> was updated successfully!`)
+          handleSuccess(
+            response,
+            a,
+            `<i>${camelCaseToRegularForm(
+              a.subEntityName
+            )} </i> was updated successfully!`
+          )
         )
         .catch(error => handleError(error, a))
     );
@@ -654,7 +759,10 @@ export const DeleteSubEntity = (action$, store) =>
       subEntityName: getSelectedEntityName(store.getState())
     }))
     .map(a => {
-      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest('delete', 'subEntity');
+      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest(
+        'delete',
+        'subEntity'
+      );
       a.sdkCall.data = {
         listId: a.entityId,
         listItemKey: a.subEntityId
@@ -667,7 +775,9 @@ export const DeleteSubEntity = (action$, store) =>
           handleSuccess(
             response,
             a,
-            `<i>${camelCaseToRegularForm(a.selectedEntity.get('name'))}</i> was deleted successfully!`
+            `<i>${camelCaseToRegularForm(
+              a.selectedEntity.get('name')
+            )}</i> was deleted successfully!`
           )
         )
         .catch(error => handleError(error, a))
@@ -675,7 +785,9 @@ export const DeleteSubEntity = (action$, store) =>
 
 export const ExecuteConfirmationDialogCallback = (action$, store) =>
   action$.ofType('EXECUTE_CONFIRM_CALLBACK').switchMap(action => {
-    const currentConfirmationModal = getConfirmationDialogType(store.getState());
+    const currentConfirmationModal = getConfirmationDialogType(
+      store.getState()
+    );
     const metaData = getConfirmationDialogMetaData(store.getState());
     switch (currentConfirmationModal) {
       case MODALS.CONFIRM_ENTITY_ACTIVE_TOGGLE:
@@ -691,7 +803,11 @@ export const ExecuteConfirmationDialogCallback = (action$, store) =>
       case MODALS.CONFIRM_ENTITY_CSV_UPLOAD:
         return [
           uploadCsv(metaData),
-          setEntityUpdating(getCurrentEntity(store.getState()), getSelectedEntityId(store.getState()), true),
+          setEntityUpdating(
+            getCurrentEntity(store.getState()),
+            getSelectedEntityId(store.getState()),
+            true
+          ),
           {
             type: 'SET_CONFIRMATION_DIALOG',
             modalType: undefined,
@@ -713,7 +829,10 @@ export const UploadCsv = (action$, store) =>
       selectedEntityId: getSelectedEntityId(store.getState())
     }))
     .map(a => {
-      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest('upload', 'singleMainEntity');
+      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest(
+        'upload',
+        'singleMainEntity'
+      );
       a.sdkCall.data = {
         [`${removeLastLetter(a.entityName)}Id`]: a.selectedEntityId,
         file: a.target
@@ -737,7 +856,9 @@ export const UploadCsv = (action$, store) =>
           }
 
           if (response.result.numberOfFailures > 0) {
-            Toast.error(`Number of Failures: ${response.result.numberOfFailures}`);
+            Toast.error(
+              `Number of Failures: ${response.result.numberOfFailures}`
+            );
           }
         })
         .map(response => ({
@@ -758,7 +879,10 @@ export const DownloadCsv = (action$, store) =>
       selectedEntityName: getSelectedEntityName(store.getState())
     }))
     .map(a => {
-      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest('download', 'singleMainEntity');
+      a.sdkCall = entitiesMetaData[a.entityName].entityApiRequest(
+        'download',
+        'singleMainEntity'
+      );
       a.sdkCall.data = {
         [`${removeLastLetter(a.entityName)}Id`]: a.selectedEntityId
       };